Xresources: fix url grab to work when /bin/sh is dash
When called like "/bin/sh <path/to/script> <args...>" the script is
run like dash. Run the script instead as an executable by using the -c
flag, which causes the bang in xterm-find-urls-dmenu.sh to be heeded.

The need for /bin/sh to start with is so we can use the HOME
environment variable for the path.